delete TopHangman::Board, move states of hangman to TopHangman::Rener
